requirements.txt
# Runtime dependencies for beerbotplus
# Install with: py -3 -m pip install -r requirements.txt

discord.py==2.6.4
# tensorflow-cpu>=2.12.0  # optional: very large; include only if you need TensorFlow

# Dev/test dependencies can go in requirements-dev.txt
